I may have done it

+++ CuraEngine/src/infill.h
@@ -11,7 +11,7 @@ void generateAutomaticInfill(const Polygons& in_outline, Polygons& result, int e
 void generateGridInfill(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, int infillOverlap, double rotation);
 void generateCubeInfill(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, int infillOverlap, double rotation,int offset);
 void generateLineInfill(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, int infillOverlap, double rotation);
-void generateCubeLineInfill(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, int infillOverlap, double rotation, double offset);
+void generateCubeLineInfill(const Polygons& in_outline, Polygons& result, int extrusionWidth, double lineSpacing, int infillOverlap, double rotation, double offset);
 
 }//namespace cura
